[2865116] Status-Bar message when doing a "Run Command" now correctly states "Password Copied".
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/trunk@2907 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ pwsafe/pwsafe/src/ui/Windows/AddEdit_Basic.cpp
@@ -893,7 +893,7 @@ void CAddEdit_Basic::OnSendEmail()
   StringX sx_email = StringX(M_email());
 
   M_pDbx()->SendEmail(sx_email.c_str());
-  M_pDbx()->UpdateLastClipboardAction((int)CItemData::EMAIL);
+  M_pDbx()->UpdateLastClipboardAction(CItemData::EMAIL);
 }
 
 LRESULT CAddEdit_Basic::OnCallExternalEditor(WPARAM, LPARAM)
